requirements-lock.txt
django
djangorestframework
channels
channels_redis